+++ CHANGELOG.md
@@ -19,6 +19,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Security
 
+## @meteora-ag/dlmm [1.0.52] - PR #90
+
+### Added
+
+- extra field (`endPrice`) in swapQuote
+
 ## @meteora-ag/dlmm [1.0.51] - PR #94
 
 ### Added